[skip ci] refinement (#50)
