+++ docs/requirements.txt
@@ -3,7 +3,7 @@ nbsphinx>=0.8.6,<0.8.7
 pyenchant>=3.1.1
 pypandoc
 requests
-sphinx>=4.1.2
+sphinx>=4,<4.0.3
 sphinx-autodoc-annotation
 sphinx-autodoc-typehints>=1.10.3
 sphinx-click>=3.0.1
