--- tests/test_cna.py
@@ -1,7 +1,0 @@
-from ._helpers import skip_mpi4py
-
-
-class TestCNA:
-    @skip_mpi4py
-    def test_infercna(self):
-        assert 1 == 1
