+++ tests/test_datasets.py
@@ -1,7 +1,25 @@
 import trisicell as tsc
 
+from ._helpers import skip_rpy2
+
 
 class TestDatasets:
+    @skip_rpy2
+    def test_simulate_1(self):
+        df_in = tsc.datasets.simulate(
+            n_cells=100, n_muts=100, n_clones=5, alpha=0.001, beta=0.4, missing=0.2
+        )
+        is_cf = tsc.ul.is_conflict_free_gusfield(df_in)
+        assert not is_cf
+
+    @skip_rpy2
+    def test_simulate_2(self):
+        df_ground = tsc.datasets.simulate(
+            n_cells=100, n_muts=100, n_clones=5, alpha=0, beta=0, missing=0
+        )
+        df_noisy = tsc.datasets.add_noise(df_ground, alpha=0.001, beta=0.4, missing=0.2)
+        assert not tsc.ul.is_conflict_free_gusfield(df_noisy)
+
     def test_load_datasets(self):
         # TODO: complete
         adata = tsc.datasets.acute_lymphocytic_leukemia1()
