+++ trisicell/pp/__init__.py
@@ -6,11 +6,21 @@ from trisicell.pp._binary import (
     binarym_filter_nonsense_mutations,
     binarym_filter_private_mutations,
     binarym_statistics,
+    consensus_combine,
 )
-from trisicell.pp._consensus import consensus_combine
 from trisicell.pp._readcount import (
-    add_vaf_to_layers,
     build_scmatrix,
+    filter_mut_mutant_must_present_in_at_least,
+    filter_mut_reference_must_present_in_at_least,
+    filter_mut_vaf_greater_than_coverage_mutant_greater_than,
+    filter_snpeff,
+    keep_cell_by_list,
+    keep_mut_by_list,
+    remove_cell_by_list,
+    remove_mut_by_list,
+    statistics,
+)
+from trisicell.pp._tmp import (
     bulk_total_mutant_vaf,
     calc_relative_expression,
     count_mut_in_a_gene,
@@ -18,26 +28,17 @@ from trisicell.pp._readcount import (
     filter_mut_by_confidence_score,
     filter_mut_coverage_mutant_must_be_greater_than,
     filter_mut_mutant_in_groups_must_present_in_at_least,
-    filter_mut_mutant_must_present_in_at_least,
     filter_mut_reference_in_groups_must_present_in_at_least,
-    filter_mut_reference_must_present_in_at_least,
     filter_mut_total_coverage_must_be_greater_than,
-    filter_mut_vaf_greater_than_coverage_mutant_greater_than,
     get_germline_variants,
     get_vaf,
     get_vaf_with_list,
-    keep_cell_by_list,
-    keep_mut_by_list,
     local_cluster_cells_then_merge_muts_pseudo_bulk,
     merge_cells_using,
     most_frequent_muts,
     muts_in_a_subset_of_cells,
     n_muts_per_cell_with_vaf,
-    remove_cell_by_list,
     remove_indels,
-    remove_mut_by_list,
-    statistics,
-    to_dendro,
 )
 from trisicell.pp._tree import collapse
 
@@ -48,8 +49,17 @@ __all__ = (
     binarym_filter_private_mutations,
     binarym_statistics,
     consensus_combine,
-    add_vaf_to_layers,
     build_scmatrix,
+    filter_mut_mutant_must_present_in_at_least,
+    filter_mut_reference_must_present_in_at_least,
+    filter_mut_vaf_greater_than_coverage_mutant_greater_than,
+    filter_snpeff,
+    keep_cell_by_list,
+    keep_mut_by_list,
+    remove_cell_by_list,
+    remove_mut_by_list,
+    statistics,
+    collapse,
     bulk_total_mutant_vaf,
     calc_relative_expression,
     count_mut_in_a_gene,
@@ -57,25 +67,15 @@ __all__ = (
     filter_mut_by_confidence_score,
     filter_mut_coverage_mutant_must_be_greater_than,
     filter_mut_mutant_in_groups_must_present_in_at_least,
-    filter_mut_mutant_must_present_in_at_least,
     filter_mut_reference_in_groups_must_present_in_at_least,
-    filter_mut_reference_must_present_in_at_least,
     filter_mut_total_coverage_must_be_greater_than,
-    filter_mut_vaf_greater_than_coverage_mutant_greater_than,
     get_germline_variants,
     get_vaf,
     get_vaf_with_list,
-    keep_cell_by_list,
-    keep_mut_by_list,
     local_cluster_cells_then_merge_muts_pseudo_bulk,
     merge_cells_using,
     most_frequent_muts,
     muts_in_a_subset_of_cells,
     n_muts_per_cell_with_vaf,
-    remove_cell_by_list,
     remove_indels,
-    remove_mut_by_list,
-    statistics,
-    to_dendro,
-    collapse,
 )
