+++ trisicell/tl/score/__init__.py
@@ -1,1 +1,1 @@
-from trisicell.tl.score._score import ad, cc, dl, mltd, tpted
+from trisicell.tl.score._ours import ad, cc, dl, mltd, tpted
