+++ trisicell/tl/score/_others.py
@@ -1,0 +1,18 @@
+def bourque():
+    # TODO: implement
+    return None
+
+
+def mp3():
+    # TODO: implement
+    return None
+
+
+def disk():
+    # TODO: implement
+    return None
+
+
+def caset():
+    # TODO: implement
+    return None
