+++ trisicell/ul/__init__.py
@@ -13,8 +13,18 @@ from trisicell.ul._packages import (
     import_mpi4py,
     import_rpy2,
 )
+from trisicell.ul._tmp import (
+    colorize_one,
+    colorize_two,
+    difference_between,
+    flips_in_cells,
+    get_database,
+    split_mut,
+    translate_gene,
+)
 from trisicell.ul._trees import (
     cells_rooted_at,
+    is_leaf,
     muts_rooted_at,
     partition_cells,
     root_id,
@@ -25,14 +35,9 @@ from trisicell.ul._trees import (
 from trisicell.ul._utils import (
     calc_score_tree,
     cleanup,
-    colorize_one,
-    colorize_two,
     count_flips,
-    difference_between,
     dir_base,
     dirbase,
-    flips_in_cells,
-    get_database,
     get_file,
     get_param,
     infer_rates,
@@ -43,14 +48,12 @@ from trisicell.ul._utils import (
     log_output,
     mkdir,
     remove,
-    split_mut,
     stat,
     timeit,
     tmpdir,
     tmpdirsys,
     tmpfile,
     tqdm_joblib,
-    translate_gene,
     with_timeout,
 )
 
@@ -100,4 +103,5 @@ __all__ = (
     tqdm_joblib,
     translate_gene,
     with_timeout,
+    is_leaf,
 )
